added unicode support for fonts
This will make emacs capable of displaying Emojis. Emojis support are
good because switching into an Erc-buffer where people had posted
Emojies, or a README.md which contains Emojis, the buffer would slow
down everything, and the characters would not get display anyways. Now
buffer switching is snappy and the characters are displayed.

This require the following fonts to be installed on the system:

  * [DejaVu Sans and DejaVu Sans Mono][dejavu]
  * [Quivira][quivira]
  * [Noto Sans and Noto Sans Symbol][noto]
  * [Symbola][symbola] by George Douros

[dejavu]: http://dejavu-fonts.org/wiki/Download
[quivira]: http://www.quivira-font.com/downloads.php
[noto]: http://www.google.com/get/noto/
[symbola]: http://www.fontspace.com/unicode-fonts-for-ancient-scripts/symbola