Remap page up and down to define and execute macros
These keys are far away from the home-row on my Saber68 keyboard, and
normally F3 and F4 are used to define and execute keyboard macros:
There are no function keys on the Saber68[1], so I use PgUp and PgDn
for this purpose instead--when a macro is defined it usually needs to
be run a couple of times, so tapping PgDn with no modifier is nice.

I love my keyboard.

[1]: Well, they are there if one uses the Fn key plus the numbers.